use `Kernel#Array` to cast values to an Array
squash out error prone local logic for array casting, use `Kernel#Array`
instead.

the older logic would fail to cast in the case that the value was a String,
requiring a tight integration between the forms input objects and the
application's forms/models. since the very next line is going to fail if we fail
to cast to an Array, let's do so aggressively here.

it looks as though this line once used `Array#wrap` for this, but this only
works if the object responds to `#to_ary` (Ruby's implicit cast); otherwise
it results in a value like `[["blah"]]`. `Kernel#Array` is willing to perform an
explicit cast.